+++ tower/src/util/mod.rs
@@ -20,7 +20,7 @@ mod then;
 
 pub use self::{
     and_then::{AndThen, AndThenLayer},
-    boxed::{BoxService, UnsyncBoxService},
+    boxed::{BoxLayer, BoxService, UnsyncBoxService},
     either::Either,
     future_service::{future_service, FutureService},
     map_err::{MapErr, MapErrLayer},