+++ tower/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Unreleased
 
 - **builder**: Add `ServiceBuilder::map_future` for transforming the futures produced
-  by a service.
+- **util**: Add `BoxLayer` for creating boxed `Layer` trait objects.
 
 # 0.4.5 (February 10, 2021)